# docs/swagger.json
{
  "definitions": {},
  "info": {
    "title": "node-todo",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3)."
  },
  "paths": {
    "/api/v1/todos": {
      "get": {
        "summary": "/api/v1/todos",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {}
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "clientError"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "code": "validationError",
                      "detail": "Unknown field(s)"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ],
        "produces": [
          "application/json"
        ]
      },
      "post": {
        "summary": "/api/v1/todos",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "ead1840e-f010-40c0-9246-273326a877a0"
                },
                "title": {
                  "type": "string",
                  "example": "w`H9isQcQz3])U#$-Ks^&MqY;32f$]'VA>E.hCfVSB`>VAuaeOz)n(\\/:^n_<u"
                },
                "description": {
                  "type": "string",
                  "example": ")Ss5%\\B"
                },
                "state": {
                  "type": "string",
                  "example": "created"
                },
                "created": {
                  "type": "string",
                  "example": "2023-06-20 16:19:18"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "clientError"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      },
                      "attr": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "code": "validationError",
                      "detail": "This field is required",
                      "attr": "title"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/api/v1/todos/{id}": {
      "get": {
        "summary": "/api/v1/todos/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "9cb754cf-f9cb-43bc-ab34-5c31a2edc3c5"
                },
                "title": {
                  "type": "string",
                  "example": "M18,Wgsr6+"
                },
                "description": {
                  "type": "string",
                  "example": "y[X,ei%1QG"
                },
                "state": {
                  "type": "string",
                  "example": "created"
                },
                "created": {
                  "type": "string",
                  "example": "2023-06-20 16:19:18"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "clientError"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "code": "notFound",
                      "detail": "Not Found"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ],
        "produces": [
          "application/json"
        ]
      },
      "patch": {
        "summary": "/api/v1/todos/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "91e22994-01c0-4398-b58f-aace78db34a1"
                },
                "title": {
                  "type": "string",
                  "example": "vF[acWdaYB"
                },
                "description": {
                  "type": "string",
                  "example": "ed+;)^V#ew"
                },
                "state": {
                  "type": "string",
                  "example": "inProgress"
                },
                "created": {
                  "type": "string",
                  "example": "2023-06-20 16:19:18"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "clientError"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "code": "notFound",
                      "detail": "Not Found"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ],
        "produces": [
          "application/json"
        ]
      },
      "delete": {
        "summary": "/api/v1/todos/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {},
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "clientError"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "code": "notFound",
                      "detail": "Not Found"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ],
        "produces": [
          "application/json"
        ]
      }
    }
  },
  "swagger": "2.0",
  "tags": [
    {
      "name": "todos"
    }
  ]
}